feat: Add serializers for UI views (#31)
This PR adds serializers for the UI views we will add in a future PR.
These views intentionally expose only what's necessary for the frontend
views. We will create new serializers should we need a different view
(e.g., for a more traditional REST API not just hydrating the frontend).