feat: Use slack for fetching user name and avatar (#42)
Turns out this is not easily achievable with IAP as previously thought.
I tried out using the google people api, but this proved to be a heavier
lift than just leveraging our existing Slack connection. This slack sync
only happens once on account creation and in the future we may want to
implement a cron job to do this sync daily or something, but for now
I've added a User model action in the django admin to resync.